+++ Tests/CryptoSwiftTests/AESTests.swift
@@ -363,34 +363,9 @@ final class AESTests: XCTestCase {
     }
 }
 
-#if !CI
-
-    extension AESTests {
-        func testAESEncryptPerformance() {
-            let key: Array<UInt8> = [0x2b, 0x7e, 0x15, 0x16, 0x28, 0xae, 0xd2, 0xa6, 0xab, 0xf7, 0x15, 0x88, 0x09, 0xcf, 0x4f, 0x3c]
-            let iv: Array<UInt8> = [0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0a, 0x0b, 0x0c, 0x0d, 0x0e, 0x0f]
-            let message = Array<UInt8>(repeating: 7, count: 1024 * 1024)
-            let aes = try! AES(key: key, blockMode: .CBC(iv: iv), padding: .pkcs7)
-            measure {
-                _ = try! aes.encrypt(message)
-            }
-        }
-
-        func testAESDecryptPerformance() {
-            let key: Array<UInt8> = [0x2b, 0x7e, 0x15, 0x16, 0x28, 0xae, 0xd2, 0xa6, 0xab, 0xf7, 0x15, 0x88, 0x09, 0xcf, 0x4f, 0x3c]
-            let iv: Array<UInt8> = [0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0a, 0x0b, 0x0c, 0x0d, 0x0e, 0x0f]
-            let message = Array<UInt8>(repeating: 7, count: 1024 * 1024)
-            let aes = try! AES(key: key, blockMode: .CBC(iv: iv), padding: .pkcs7)
-            measure {
-                _ = try! aes.decrypt(message)
-            }
-        }
-    }
-#endif
-
 extension AESTests {
     static func allTests() -> [(String, (AESTests) -> () -> Void)] {
-        var tests = [
+        let tests = [
             ("testAESEncrypt", testAESEncrypt),
             ("testAESEncrypt2", testAESEncrypt2),
             ("testAESEncrypt3", testAESEncrypt3),
@@ -414,13 +389,6 @@ extension AESTests {
             ("testIssue411", testIssue411),
             ("testAESWithWrongKey", testAESWithWrongKey),
         ]
-
-        #if !CI
-            tests += [
-                ("testAESEncryptPerformance", testAESEncryptPerformance),
-                ("testAESDecryptPerformance", testAESDecryptPerformance),
-            ]
-        #endif
         return tests
     }
 }
