+++ Tests/CryptoSwiftTests/ChaCha20Tests.swift
@@ -108,37 +108,14 @@ final class ChaCha20Tests: XCTestCase {
     }
 }
 
-#if !CI
-
-    extension ChaCha20Tests {
-        func testChaCha20Performance() {
-            let key: Array<UInt8> = [0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0a, 0x0b, 0x0c, 0x0d, 0x0e, 0x0f, 0x10, 0x11, 0x12, 0x13, 0x14, 0x15, 0x16, 0x17, 0x18, 0x19, 0x1a, 0x1b, 0x1c, 0x1d, 0x1e, 0x1f]
-            let iv: Array<UInt8> = [0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07]
-            let message = Array<UInt8>(repeating: 7, count: 1024 * 1024)
-            measureMetrics([XCTPerformanceMetric.wallClockTime], automaticallyStartMeasuring: true, for: { () -> Void in
-                do {
-                    _ = try ChaCha20(key: key, iv: iv).encrypt(message)
-                } catch {
-                    XCTFail()
-                }
-                self.stopMeasuring()
-            })
-        }
-    }
-#endif
-
 extension ChaCha20Tests {
     static func allTests() -> [(String, (ChaCha20Tests) -> () -> Void)] {
-        var tests = [
+        let tests = [
             ("testChaCha20", testChaCha20),
             ("testCore", testCore),
             ("testVector1Py", testVector1Py),
             ("testChaCha20EncryptPartial", testChaCha20EncryptPartial),
         ]
-
-        #if !CI
-            tests += [("testChaCha20Performance", testChaCha20Performance)]
-        #endif
 
         return tests
     }
