+++ Tests/CryptoSwiftTests/Default-568h@2x.png
